If the dimensions of options aren't the same as the dataset, it won't do the options. Changed BoundedMDScan 's arguments to take blockDimensions

+++ src/main/scala/org/hdfgroup/spark/hdf5/DefaultSource.scala
@@ -28,7 +28,7 @@ class DefaultSource extends RelationProvider {
     val extensions = parameters.getOrElse("extension", "h5").split(",").map(_.trim)
     val chunkSize = parameters.getOrElse("window size", "10000").toInt
     val start = parameters.getOrElse("start", "-1").split(",").map(_.toLong).distinct
-    val block = parameters.getOrElse("block", "-1").split(",").map(_.toLong).distinct
+    val block = parameters.getOrElse("block", "-1").split(",").map(_.toInt).distinct
 
     new HDF5Relation(paths, dataset, extensions, chunkSize, start, block)(sqlContext)
   }
